% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Alpha_Div.R
\name{One_Alpha_Div}
\alias{One_Alpha_Div}
\title{A convenient function for calculating alpha diversity}
\usage{
One_Alpha_Div(otu, tree = NULL, rrarefy = T, melt = F)
}
\arguments{
\item{otu}{The OTUs table, which row are OTUs and column are samples}

\item{tree}{Tree file}

\item{rrarefy}{Logical, whether to rrarefy}

\item{melt}{Logical, whether to melt the result according to the alpha div index}
}
\value{
A data.frame
}
\description{
A convenient function for calculating alpha diversity
}
\examples{
One_Alpha_Div(Protist$POTUs)

}
