% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Contrast.R
\name{One_EdgeR}
\alias{One_EdgeR}
\title{A convenient way to do one-way edgeR analysis}
\usage{
One_EdgeR(otu, design, treatment, sig = 0.05, padj = "BH", prefix = NULL)
}
\arguments{
\item{otu}{The OTUs table, which row are OTUs and column are samples}

\item{design}{The metadata file, a data.frame, should have column of sampleid and group information}

\item{treatment}{The column name of the treatment id}

\item{sig}{Numeric, the threshold of significant p-value}

\item{padj}{Method for p adjust}

\item{prefix}{Prefix for files}
}
\value{
A data.frame and files
}
\description{
For loop was used to conduct the multi-group pairwise edgeR analysis
}
\examples{
otu <- Protist$POTUs
design <- Protist$PDesign
One_EdgeR(otu,design = design,treatment = 'precipitation',prefix = 'prec')
One_EdgeR(otu,design = design,treatment = 'temperature',prefix = 'temp')


}
