% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Tax_Abundance.R
\name{One_Filter_Aabundance}
\alias{One_Filter_Aabundance}
\title{A method for screening OTU tables according to abundance}
\usage{
One_Filter_Aabundance(otu, input_format = "count", mt = NULL, num = NULL)
}
\arguments{
\item{otu}{OTUs table, which row are OTUs(or other ranks) and column are samples}

\item{input_format}{The value of the OTUs table, counts or relative (abundance)}

\item{mt}{The threshold of mean relative abundance}

\item{num}{The numbers of OTUs you want to have in the result}
}
\value{
A data.frame with relative abundance
}
\description{
OTUs tables can be filtered based on abundance thresholds or number thresholds
}
\examples{
otu <- Protist$POTUs
One_Filter_Aabundance(otu,mt=0.001)
One_Filter_Aabundance(otu,num=200)
One_Filter_Aabundance(otu,mt=0.01,num = 3)
One_Filter_Aabundance(otu,mt=0.01,num = 40)
## Filter Abundance table with other ranks
tax <- Protist$PTAXs
res_phy <- One_Tax_Abundance(tax=tax,otu=otu,tar='phylum',norm=TRUE)
One_Filter_Aabundance(res_phy,mt=0.01,num = 20)

}
