% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Format2Veen.R
\name{One_Format2Veen}
\alias{One_Format2Veen}
\title{An easy way to transform data for Veen and Upset plot}
\usage{
One_Format2Veen(otu, design, group)
}
\arguments{
\item{otu}{The OTUs table, which row are OTUs and column are samples}

\item{design}{The design file, a data.frame, should have column of sampleid and group information}

\item{group}{The column name of the group}
}
\value{
A data.frame
}
\description{
Transform OTUs table according to the design table
}
\examples{
otu <- Protist$POTUs
design <- Protist$PDesign
group='trt'
One_Format2Veen(otu,design,'trt')
}
