% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Tax_Abundance.R
\name{One_Group_Abundance}
\alias{One_Group_Abundance}
\title{Integrating abundance according to the group information}
\usage{
One_Group_Abundance(otu, metadata, id = "row.names", group = NULL, norm = T)
}
\arguments{
\item{otu}{OTUs table, which row are OTUs(or other ranks) and column are samples}

\item{metadata}{The metadata file, a data.frame, should have column of sampleid and group information}

\item{id}{The sample id, which could also be found in OTUs table}

\item{group}{The column name of the group}

\item{norm}{Logical, if the counts need to be normlized}
}
\value{
A data.frame with 3 columns
}
\description{
You can generate counts table or relative abundance table according to the treatments
}
\examples{
otu <- Protist$POTUs
design <- Protist$PDesign
One_Group_Abundance(otu,metadata=design,id='row.names',group = 'temperature',norm = FALSE)
One_Group_Abundance(otu,metadata=design,id='row.names',group = 'temperature',norm = TRUE)
tax <- Protist$PTAXs
res_phy <- One_Tax_Abundance(tax=tax,otu=otu,tar='phylum',norm=TRUE)
One_Group_Abundance(otu=res_phy,metadata=design,id='row.names',group = 'temperature',norm = FALSE)


}
