% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_LSD.R
\name{ONE_LSD}
\alias{ONE_LSD}
\title{One step for multi-group LSD Test}
\usage{
ONE_LSD(data = ARGs, group = "gene", compare = "trt", value = "value")
}
\arguments{
\item{data}{Whole data.frame}

\item{group}{Column containing grouping information}

\item{compare}{Column containing the difference analysis group}

\item{value}{Column containing the value}
}
\value{
A data.frame containing mean, std, results of LSD
}
\description{
A great way to test the significance if your data possess multi-groups
}
\examples{
data(ARGs)
ONE_LSD(data=ARGs,group='gene',compare='trt',value='value')

}
