% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_MRPP.R
\name{One_MRPP}
\alias{One_MRPP}
\title{One step for MRPP and pairwise comparison}
\usage{
One_MRPP(
  otu,
  metadata,
  id = "row.names",
  gpid = "trt",
  dis_method = "bray",
  perm = 999,
  p.adj = "BH",
  prefix = NULL
)
}
\arguments{
\item{otu}{The OTUs table, which row are OTUs and column are samples}

\item{metadata}{The metadata file, a data.frame, should have column of sampleid and group information}

\item{id}{The sample id, which could also be found in OTUs table}

\item{gpid}{The column name of the group id}

\item{dis_method}{The method to construct the distance matrix, including "manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao" ,"mahalanobis".}

\item{perm}{The number of permutations times}

\item{p.adj}{The method to adjust the p-value, including 'bonferroni','holm','hochberg','hommel','fdr','BH','BY','none'}

\item{prefix}{The prefix of the output file}
}
\value{
A list of MRPP result
}
\description{
Conveniently carried out MRPP and pairwise comparison, the result would be writeen in directory  "MRPP result"
}
\examples{
One_MRPP(otu=Protist$POTUs,metadata=Protist$PDesign,id='row.names',gpid = 'trt',dis_method='bray',perm=999,p.adj='BH')
}
