% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_SigLetters.R
\name{One_SigLetters}
\alias{One_SigLetters}
\title{Add letters according to the p-values}
\usage{
One_SigLetters(df, id, sort = NULL, pvalue)
}
\arguments{
\item{df}{A data.frame with group and mean/median information}

\item{id}{The column of group information}

\item{sort}{The column of mean/median values}

\item{pvalue}{The p-value matrix, both n*n matrix and n*3 matrix can be accepted}
}
\value{
A data.frame with significance letter result
}
\description{
A good way to add letters to results of  multiple comparisons
}
\examples{
df <- data.frame(id=c('A','B','C','D'),mean=c(0.6,0.4,0.7,0.1))
pvalue <- data.frame(cp1 = c('A','A','A','B','B','C'),cp2=c('B','C','D','C','D','D'),p =c('0.09','0.1','0.04','0.03','0.045','0.01'))
One_SigLetters(df,id="id",sort='mean',pvalue = pvalue)

}
