% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Spearman_Net.R
\name{One_Spearman_Net}
\alias{One_Spearman_Net}
\title{One step to construct Spearman Network}
\usage{
One_Spearman_Net(
  normtab,
  tax,
  rank = "OTU",
  abundance = 0,
  appear = 0,
  num = NULL,
  cor_type = "spearman",
  P.adj = "BH",
  rt = 0.7,
  pt = 0.05,
  write = F,
  prefix = NULL
)
}
\arguments{
\item{normtab}{Standardized OTU table, which row are OTUs and column are samples}

\item{tax}{TAXs table, which row are OTUs and column are ranks}

\item{rank}{Which rank of the normtab, usually "OTU"}

\item{abundance}{Abundance threshold, numeric}

\item{appear}{The least occurrence of OTUs in the sample, numeric}

\item{num}{The numbers of OTUs you want to keep to construct the network, numeric}

\item{cor_type}{Types of correlation, could be set as 'pearson','kendall','spearman'}

\item{P.adj}{Method for p adjust, could be set as 'bonferroni','holm','hochberg','hommel','fdr','BH','BY','none'}

\item{rt}{The threshold of correlation, numeric}

\item{pt}{The threshold of p-value, numeric}

\item{write}{Logical, whether to write result to files for Gephi or Cytoscape}

\item{prefix}{Prefix for files}
}
\value{
A list of igraph object, node table, edge table, and correlation matrix
}
\description{
A easy way to construct Spearman Network and caculate the network features altogether
}
\examples{
normtab <- Protist$POTUs
normtab <- as.data.frame(t(t(normtab)/colSums(normtab)))
tax <- Protist$PTAXs
One_Spearman_Net(normtab=normtab,tax=tax,rank='OTU',abundance = 0.001,appear=4,num=NULL,cor_type = 'spearman',P.adj = 'BH',rt=0.7,pt=0.05,write = TRUE)
}
