% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Tax_Abundance.R
\name{One_Tax_Abundance}
\alias{One_Tax_Abundance}
\title{A convenient way to integrate OTUs tables}
\usage{
One_Tax_Abundance(tax, otu, tar = NULL, all = F, norm = F)
}
\arguments{
\item{tax}{TAXs table, which row are OTUs and column are ranks}

\item{otu}{OTUs table, which row are OTUs and column are samples}

\item{tar}{A column name of the TAXs table}

\item{all}{Logical, if all ranks would be caculated}

\item{norm}{Logical, if the counts need to be normlized}
}
\value{
A data.frame or list
}
\description{
You can generate counts table or relative abundance table on other ranks, not just OTUs levels
}
\examples{
tax <- Protist$PTAXs
otu <- Protist$POTUs
One_Tax_Abundance(tax=tax,otu=otu,tar='phylum',norm=TRUE) ## specified target rank with normalization
One_Tax_Abundance(tax=tax,otu=otu,all=TRUE)[['phylum']] ## All ranks
One_Tax_Abundance(tax=tax,otu=otu,all=TRUE,norm = TRUE )[['phylum']] ## All ranks with normalization


}
