% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Tukey_HSD.R
\name{One_Tukey_HSD1}
\alias{One_Tukey_HSD1}
\title{One step for multi-group Tukey HSD Test}
\usage{
One_Tukey_HSD1(data = ARGs, group = "gene", compare = "trt", value = "value")
}
\arguments{
\item{data}{Whole data.frame}

\item{group}{Column containing grouping information}

\item{compare}{Column containing the difference analysis group}

\item{value}{Column containing the value}
}
\value{
A data.frame containing mean, std, results of Tukey HSD Test
}
\description{
A great way to test the significance if your data possess multi-groups
}
\examples{
data(ARGs)
One_Tukey_HSD1(ARGs,group='gene',compare = 'location',value = 'value')
One_Tukey_HSD2(ARGs,group='gene',compare = 'location',value = 'value')


}
