% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/One_Wilcox.R
\name{One_Wilcox}
\alias{One_Wilcox}
\title{A great way to perform multi-group Wilcox Test}
\usage{
One_Wilcox(
  data = ARGs,
  group = "gene",
  compare = "location",
  value = "value",
  sort = "mean"
)
}
\arguments{
\item{data}{Whole data.frame}

\item{group}{Column containing grouping information}

\item{compare}{Column containing the difference analysis group}

\item{value}{Column containing the value}

\item{sort}{A way to sort the data for add letters with two modes "mean" and "median"}
}
\value{
A list of results
}
\description{
A great way to perform multi-group Wilcox Test
}
\examples{
One_Wilcox(data = ARGs,group = 'gene',compare = 'location',value = 'value',sort = 'mean')
One_Wilcox(data = ARGs,group = 'gene',compare = 'location',value = 'value',sort = 'median')

}
