% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vis_map.R
\name{Vis_map}
\alias{Vis_map}
\title{To normalize data to a certin ranges for visualization}
\usage{
Vis_map(x, range = c(0, 1), from.range = NA)
}
\arguments{
\item{x}{original matrix or vector or data.frame, must be numeric}

\item{range}{your expected range}

\item{from.range}{}
}
\value{
matrix or vector or data.frame after mapping to ranges
}
\description{
A convenient function to modify the data distribution range
}
\examples{
Vis_map(matrix(1:9,ncol = 3))
Vis_map(1:9)
Vis_map(data.frame(a=1:3,b = 7:9))
}
